+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/components/PeriodDefinition/PeriodDefinition.tsx
@@ -5,14 +5,12 @@ import DEFAULT_LOCALE_EN from '../../locale';
 import { PeriodDefinitionProps } from './PeriodDefinition.types';
 
 const PeriodDefinition: FC<PeriodDefinitionProps>= ({
-    isDisplayed,
     locale = DEFAULT_LOCALE_EN,
     localeKey,
-}) => isDisplayed 
-    ? (
-        <PeriodText>
-            {locale[localeKey] || DEFAULT_LOCALE_EN[localeKey]}
-        </PeriodText>
-    ): null;
+}) => (
+    <PeriodText>
+        {locale[localeKey] || DEFAULT_LOCALE_EN[localeKey]}
+    </PeriodText>
+);
 
 export default PeriodDefinition;
