+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/components/PeriodDefinition/PeriodDefinition.types.ts
@@ -1,7 +1,6 @@
 import { Locale } from '../../types';
 
 export interface PeriodDefinitionProps {
-    isDisplayed: boolean;
     locale: Locale;
     localeKey: string;
 }
