+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/MonthsSelect/MonthsSelect.tsx
@@ -22,7 +22,6 @@ const MonthsSelect: FC<MonthsSelectProps> = ({
     return (
         <>
             <PeriodDefinition
-                isDisplayed={true}
                 locale={locale}
                 localeKey='prefixMonths'
             />
