+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/WeekDaysSection/WeekDaysSection.types.ts
@@ -11,5 +11,4 @@ export interface WeekDaysSectionProps {
     nthWeekDays: number[];
     setNthWeekDays: (nthWeekDays: number[]) => void;
     isMonthPeriodDisplayed: boolean;
-    isWeekPeriodDisplayed: boolean;
 }
