+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/WeekDaysSection/WeekDaysSelect/WeekDaysSelect.types.ts
@@ -2,5 +2,4 @@ import { FieldProps } from '#src-app/components/cron/types';
 
 export interface WeekDaysSelectProps extends FieldProps {
     humanizeLabels: boolean;
-    isWeekPeriodDisplayed: boolean;
 }
